use policy to handle permissions

+++ app/Http/Controllers/CityController.php
@@ -13,23 +13,13 @@ use App\City;
 use App\Country;
 use App\Http\Requests\CityRequest;
 use Exception;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Yajra\DataTables\DataTables;
 
 class CityController extends Controller
 {
-    function __construct()
-    {
-        $this->middleware(
-'permission:city-list|city-create|city-edit|city-delete',
-            ['only' => ['index','store']]
-        );
-        $this->middleware('permission:city-create', ['only' => ['create','store']]);
-        $this->middleware('permission:city-edit', ['only' => ['edit','update']]);
-        $this->middleware('permission:city-delete', ['only' => ['destroy']]);
-    }
-
     /**
      * Display a listing of the resource.
      *
@@ -39,6 +29,8 @@ class CityController extends Controller
      */
     public function index(Request $request)
     {
+        $this->authorize('viewAny', City::class);
+
         if ($request->ajax()) {
             $data = City::latest()->with('country');
             return Datatables::of($data)
@@ -63,9 +55,12 @@ class CityController extends Controller
      * Show the form for creating a new resource.
      *
      * @return Response
+     * @throws AuthorizationException
      */
     public function create()
     {
+        $this->authorize('create', City::class);
+
         $countries = Country::pluck("name", "id");
         return view('dashboard.cities.create', compact('countries'));
     }
@@ -73,11 +68,13 @@ class CityController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  CityRequest $request
+     * @param CityRequest $request
      * @return Response
+     * @throws AuthorizationException
      */
     public function store(CityRequest $request)
     {
+        $this->authorize('create', City::class);
         City::create($request->all());
         return redirect()->route('cities.index')
             ->with('success', 'City created successfully');
@@ -92,14 +89,17 @@ class CityController extends Controller
     {
         //
     }
+
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  City $city
+     * @param City $city
      * @return Response
+     * @throws AuthorizationException
      */
     public function edit(City $city)
     {
+        $this->authorize('update', $city);
         $countries = Country::pluck("name", "id");
         return view('dashboard.cities.edit', compact('city', 'countries'));
     }
@@ -110,9 +110,11 @@ class CityController extends Controller
      * @param CityRequest $request
      * @param City $city
      * @return Response
+     * @throws AuthorizationException
      */
     public function update(CityRequest $request, City $city)
     {
+        $this->authorize('update', $city);
         $city->update($request->all());
 
         return redirect()->route('cities.index')
@@ -128,6 +130,7 @@ class CityController extends Controller
      */
     public function destroy(City $city)
     {
+        $this->authorize('delete', $city);
         $city->delete();
         return redirect()->route('cities.index')
             ->with('error', 'City Deleted successfully');
